make_and_checkout_output_branch is different for split out/in
